Fix issue re-rendering table cells when the table becomes valid (#1041)
# Pull Request

## 🤨 Rationale

Fixes #1039 

## 👩‍💻 Implementation

When an invalid table become valid again, all the cells are re-added to
the DOM. As a result, the `connectedCallback` is called again for a cell
that had previously been connected. Calling the `connectedCallback`
again caused the cell template to be added an additional time to the
`cellContentContainer`. To fix the problem, we need to clean up the cell
template in the `disconnectedCallback`.

## 🧪 Testing

- Added new unit test that failed prior to my fix
- Manually tested in storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.