Add new save icon and update icon import instructions (#1037)
# Pull Request

## 🤨 Rationale

Fixes #1035. Brandon and I also realized that our process for importing
icons relied on Illustrator access which most devs (and some designers)
don't have, so the process needed to be updated.

## 👩‍💻 Implementation

Updated the process and followed it. I had to manually modify the SVG
that Brandon provided to ensure it passed our autotest; I copied the
structure from other icons to do this.

## 🧪 Testing

New icon appears in storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.