Fix background color of dropdown listbox in color theme (#1029)
## 🤨 Rationale

Fixes #1027 

I also addressed the fact that menus/dropdowns have transparent borders
which look bad when overlaying other visuals.

## 👩‍💻 Implementation

1. I followed the example of the menu. The color theme now just overlays
a semi-transparent white over the solid background of the dropdown
listbox.
2. I also noticed that the menu/dropdown border was transparent, too.
This looks bad when overlaying background visuals (see the select
component's matrix Storybook for examples). This was easily fixed by
removing the `background-clip` value.

I had originally gone the route of making a theme-aware, opaque token
for the dropdown/menu background, but this was later rejected because it
would break our ability to change the color of the color theme. We want
to be able to change one (or a couple) tokens from the current value of
ForestGreen to something else to achieve arbitrary monochromatic looks.
Thus we do not want to bake that color into other tokens (i.e. the
combination of that color with a semi-transparent white).

## 🧪 Testing

Tested manually in Storybook

Before: (note the borders)

![image](https://user-images.githubusercontent.com/7282195/217064263-cc91c00e-da7c-496e-a13e-98ae30abb769.png)

After:

![image](https://user-images.githubusercontent.com/7282195/217102282-23ac8bff-e286-4c7e-a897-25000a0a23c0.png)


## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.