Table styling update: Headers should appear to overhang above the scrollbar (#1050)
# Pull Request

## 🤨 Rationale

Followup PR for #865 - after feedback from Nate/Rick, we want the header
row to appear to overhang above the table scroll bar.

See https://github.com/ni/nimble/issues/865#issuecomment-1410372829 for
context and screenshots.

## 👩‍💻 Implementation

Instead of setting `margin-right` on the header container based on the
scrollbar width (which left a transparent square in the top-right corner
of the table bounds above the scrollbar), add an empty `div` to the end
of the table container with that same width. Net result is that
upper-right square appears with the same background color as the
headers, and gets the overhang effect.

(Note: The reason why we need the margin in the first place is because
otherwise, the column headers would no longer align vertically with the
cells, once the scrollbar was showing.)

Before:  

![TableScrollOverhangBefore](https://user-images.githubusercontent.com/20709258/218572736-87795a2d-9d10-4af7-84f9-ae78eb2fcaa4.PNG)

After:  

![TableScrollOverhangAfter](https://user-images.githubusercontent.com/20709258/218573067-e447ba3a-4627-424d-86d4-950e8105c372.PNG)

## 🧪 Testing

Tested via Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.