Fix bug where pointer events were being swallowed by table row hover (#1055)
# Pull Request

## 🤨 Rationale

Mouse events (e.g. text selection) are not working correctly within the
table because the `::before` styling that is intended to make the row
background color change on hover is swallowing the pointer events.

## 👩‍💻 Implementation

Set `pointer-events: none;` on the `::before` styling within the table.

I also did a search for other `::before` and `::after` styles within the
repo where this might be a problem, but I didn't find any. Most of our
usages of `::before` and `::after` are to change the bottom border of
controls on hover and/or focus.

## 🧪 Testing

Verified text can now be selected correctly within the table.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.