Allow table columns to be marked as hidden (#1070)
# Pull Request

## 🤨 Rationale

We need to be able to mark columns in a table has hidden so that the
table can use them for operations, such as sorting, but have those
hidden columns not be included when the table is rendered.

This is required as part of #874 

## 👩‍💻 Implementation

- Add a boolean `column-hidden` attribute to the base column class
- Don't render header or cells for hidden columns
- Add Blazor support for hidden columns
- Add Angular support for hidden columns

## 🧪 Testing

- Manually tested in storybook, Angular, and Blazor
- Wrote new unit tests
- Added a hidden column to the table matrix tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.