Fix mistake in table-row unit test set up (#1077)
# Pull Request

## 🤨 Rationale

I stumbled across this mistake in the test set up. It doesn't look like
it was causing any issues in the tests, but it should be fixed
regardless.

## 👩‍💻 Implementation

Make the closing element tag match the opening tag.

## 🧪 Testing

None other than letting the pipeline run to ensure the unit tests are
still passing

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>