Create chromatic tests with an open action menu within a table (#1073)
# Pull Request

## 🤨 Rationale

Resolves #1058 

## 👩‍💻 Implementation

- Create theme-specific tests for the action menu being open in a table.
These couldn't all be combined into a single story with multiple themes
because having multiple menu buttons open at the same time is not valid.
- Update the styling of the background-color wrapper that is used within
`createFixedThemeStory`. Previously, the `div` had absolute positioning,
which meant the window's viewport didn't grow to fit the story's size
(it was always 0px). As a result, the menu button didn't know which way
to open, defaulted to opening up, and was clipped off the top of the
window. I changed the `div` to no longer be absolutely positioned while
maintaining the same appearance so that none of the other fixed-theme
stories were affected.

## 🧪 Testing

- Looked at chromatic diffs

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.