Add programmatic column sorting support to the table (#1080)
# Pull Request

## 🤨 Rationale

This is the primary implementation for #874. It adds programmatic
support (though `sort-direction` and `sort-index` attributes) to columns
to mark themselves as sorted.

## 👩‍💻 Implementation

In order to have TanStack better facilitate column sorting, I needed to
make the TanStack columns align with the nimble columns. Previously
columns in TanStack were generated by inspecting the fields within the
first data record, but now the TanStack columns are generated from the
slotted columns within the table. They are given a unique ID (the same
ID that we use for slotting header content), an accessor that aligns the
new `operandDataRecordFieldName` property on the column base class, and
a sort function that is determined by the new `sortOperation` property
on the column base class. With the TanStack columns aligned with the
nimble columns, the table can listen to changes to `sort-direction` and
`sort-index` on the columns and update the TanStack sort state.

Additional implementation notes:
- In an effort to get all the TanStack `state` options to be merged
together correctly, I was able to simplify some of the logic around
updating state.
- I made a few changes to the duplicate sort index validation, as
reflected in updates to the sort HLD

## 🧪 Testing

- A bunch of new unit tests
- Added sorted columns in storybook and matrix tests
- Verified sorting worked correctly in Angular and Blazor

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.