Remove div in icon template
The div in the icon template is display:block and responds to flow layout (i.e. line-height) this is unexpected as the top-level of the icon is configured flex and does flex box layout. This was causing the icon to shift in an inline context. This removes the div altogether to avoid that behavior