Set title attribute in text column cell (#1092)
## 🤨 Rationale

See #857

We want a tooltip on hover that shows the full text of a table cell.
This allows the user to see the full value even when it is truncated in
the cell.

## 👩‍💻 Implementation

Every table column type has to implement this themselves, so I've
implemented it for the text column. In the cell template, I set `title`
to the same expression used for the cell value. If the value is
null/undefined, the title is the placeholder text.

## 🧪 Testing

Added unit tests for the value and placeholder cases.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.