Support "." in table record field names (#1110)
# Pull Request

## 🤨 Rationale

Resolves #1108 

## 👩‍💻 Implementation

Rather than setting `accessorKey` on a TanStack column, which tries to
split keys on ".", I am now setting `accessorFn` that allows the
nimble-table to do the lookup into the record and no longer split on
".".

## 🧪 Testing

Added a new unit test that failed before I made my change and passes
with my change.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.