Batch table updates (#1099)
# Pull Request

## 🤨 Rationale

Resolves #1091 

## 👩‍💻 Implementation

Create `UpdateTracker` utility class that helps manage what updates are
needed for the table and queuing those updates so they run in batches
after an animation frame.

The one update to the table that does not go through the `UpdateTracker`
is updates to the table's data. This is because we didn't want the
nimble table to hold on to an extra copy of the data, which it would
need to do in order to apply it after an animation frame. Additionally,
the performance benefit of batching TanStack updates for a data update
did not seem as significant as performance benefits from other batching
where it is expected to receive multiple changes within a single
animation frame. For example, if the sorting changes, then it is
expected to receive multiple change notifications about the `sort-index`
and `sort-direction` on multiple columns all within a very short period
of time.

## 🧪 Testing

- Relied primarily on existing auto tests
- I noticed that we previously were not handling the case that
`action-menu-slot` changed after a column was created, so I added a new
unit test for that code path

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.