Set fixed height/width of div in tooltip matrix story (#1111)
## 🤨 Rationale

The tooltip matrix stories are randomly showing diffs. See #1106 

## 👩‍💻 Implementation

We hypothesize that the recent change to fallback fonts somehow
introduced this intermittency. One proposed idea was to set fixed
dimensions on the div that anchors the tooltip, in case that div was
being resized due to font loading and somehow resulting in inconsistent
layout at the time of the snapshot.

I also adjusted the padding so the test cases are arranged more densly.

## 🧪 Testing

We'll see if this reduces/elliminates the intermittency after being
submitted.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.