Replaces usages of DesignSystem.tagFor() with exported element tags (#1115)
# Pull Request

## 🤨 Rationale

Addresses most of #1095. We're moving to a custom Nimble API that
encapsulates from `DesignSystem.tagFor()` because it appears to be going
away in a future FAST version.

## 👩‍💻 Implementation

- Find and replace usages of tagFor except for those within each
component that are used to expose the tag
- Expose component tags for a few components that were missing it

## 🧪 Testing

Relying on pipeline.

I briefly considered adding `no-restricted-imports` lint configuration
to enforce this but it seemed like it would be incomplete to still allow
it in `index.ts` and `*.spec.ts` (not to mention a bit messy to add more
conditional configuration when [our configuration is already moderately
complex](https://github.com/ni/nimble/blob/d0454c23c82195d6b7530cbb3ff3da2ce21afabb/packages/nimble-components/src/.eslintrc.js#L95)).


## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.