Add documentation of some table column APIs in Storybook (#1105)
# Pull Request

## 🤨 Rationale

We've been adding new APIs to table columns in general and to specific
column types like text but haven't been documenting them. This handles
some of the tasks in #1044.

## 👩‍💻 Implementation

- add a new story dedicated to table columns. new story includes docs
for:
    - column order
    - column header content
    - column sorting
- other common attributes that don't merit their own story: field name
and placeholder
- add a new story for the text column type
- remove some content from the existing table story
- Also did some minor refactoring to increase code sharing

There are additional stories tracked in #1044 that I'm not tackling yet
like docs for action menu and column sizing.

## 🧪 Testing

Visual inspection of new stories

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>