Disable flaky tooltip tests (#1120)
# Pull Request

## 🤨 Rationale

The tooltip matrix visual tests are rendering differently in Firefox
from run to run. Until we can get them rendering in a consistent way, we
want to disable these so we don't have to keep accepting differences.

## 👩‍💻 Implementation

Used the `disableSnapshot` flag.

## 🧪 Testing

Will be tested by the pipeline build.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.