Update table header styling to not reserve space for sort icon (#1131)
# Pull Request

## 🤨 Rationale

Resolves #1127 

## 👩‍💻 Implementation

We no longer need a wrapper around the sort icons to reserve their space
when neither icon is displayed. Therefore, I got rid of the wrapping
`span` and moved the padding to the icon element.

Note: The visual diff is because the `span` was removed. The `span` by
default had a line height of 18px even though the icon is only 16px
tall. Now, the icon should be correctly vertically centered (though I
don't think anyone noticed it was 1px off from center previously).

## 🧪 Testing

- Updated unit tests
- Manually verified that the behavior is correct with a long, ellipsed
column header. Previously, the ellipsis in the column header would occur
in the same place regardless of whether or not the column was sorted.
Now, the location of the ellipsis in the column header is dependent on
whether or not the column is sorted.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.