Add invalid styling and update focus styling for text area component (#1119)
## 🤨 Rationale

See #486

This PR adds invalid styling for the text area **and** updates the
general focus/hover styling to match the updated design doc, i.e. only
color/thicken the bottom border rather than the entire border.

## 👩‍💻 Implementation

- Added two new attributes/properties: `error-visible` and `error-text`,
following the pattern already set by other components
- Needed to fork the FAST template to add our error exclamation mark
icon and the error text that appears underneath the text area. Forking
was necessary because the FAST text area doesn't have any slots we could
use.
- One challenge is that the exclamation icon needs to appear just inside
the top right corner of the text area, and it needs to be be offset
differently depending on whether there is a vertical scrollbar. A
vertical scrollbar can appear or disappear when the user resizes the
control (dragging the resize handle) or types in the control. I used an
`input` event handler and a `ResizeObserver` to detect when the
scrollbar visibility had changed.
- Updated the Storybook stories
- Updated the Angular and Blazor integration and tests
- Updated the non-error state focus/hover style to match the design doc.
This meant only coloring/highlighting the bottom border rather than the
full border. I did not implement the double-underline style from the
design doc, because it is still unclear if that is a valid, distinct
state.

Design doc:

![image](https://user-images.githubusercontent.com/7282195/225090335-f96975c6-e756-42bd-8c6a-10a9d8b3e0a6.png)
(Note that I checked with Brandon to confirm that we **don't** want the
weird spacing to the right of the exclamation mark as is shown in the
design doc. He agrees with the spacing that I ultimately used instead.)
Old:

![image](https://user-images.githubusercontent.com/7282195/225090512-bbf2ded3-44f8-44fd-ab6b-9673900c50be.png)
New:

![image](https://user-images.githubusercontent.com/7282195/225091106-f9744081-29d3-4d0a-b251-aff8f8bd2cb6.png)


## 🧪 Testing

Tested in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.