Nimble Table: cell view (custom element) and cell state changes (#1138)
# Pull Request

## 🤨 Rationale

Fixes #997 

## 👩‍💻 Implementation

Implementation is similar to what was outlined in the the
[HLD](https://github.com/ni/nimble/blob/main/packages/nimble-components/src/table/specs/cell-state-when-scrolling-hld.md)
([PR](https://github.com/ni/nimble/pull/1052)):
- Add a `TableCellView<TCellRecord, TColumnConfig>` class. TableCellView
instances will be used in `TableCell`s, instead of the old approach
(column-provided cellTemplate and cellStyles).
- Add `cellViewTag` property to `TableColumn` (links the column with the
associated `TableCellView` type)
- `TableCellView` defines a `focusedRecycleCallback()` method which is a
no-op by default. If a cell view contains a focusable element, and a
cell is focused, and then a scroll (row recycle) occurs, this gives the
cell view an opportunity to commit/blur/etc.
- We don't yet have a column type exercising this, but there's a new
autotest for this code path.
- If a cell action menu is open when a scroll/ row recycle occurs, the
action menu is closed (via the associated menu button).

## 🧪 Testing
- Updated existing autotests
- Added new tests for `focusedRecycleCallback` and the code that closes
action menus on scroll
- Manually tested table in Storybook + Angular + Blazor example apps

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.