Upgrade to TypeScript 4.7.4 (#1159)
# Pull Request

## 🤨 Rationale

Some of the storybook bug fixes in #1146 will likely rely on the
[TypeScript 4.7 support for defining entry points in
package.json](https://devblogs.microsoft.com/typescript/announcing-typescript-4-7/#package-json-exports-imports-and-self-referencing).
While we can't uptake those fixes yet, I want to remove a potential
blocker by moving us from TypeScript 4.6 to 4.7

## 👩‍💻 Implementation

1. Replaced all `package.json` instances of `typescript": "~4.6.4"` with
`typescript": "~4.7.4"`. That's the latest patch on the 4.7 minor
release.
2. Ran `npm install` to regenerate `package-lock.json`

We're not updating to TypeScript 4.8 or 5.0 yet because of blockers like
#810.

## 🧪 Testing

Ran `npm run pack` and brought the `nimble-angular` tgz into
SystemLinkShared. Verified that it still builds and the example app runs
even though that project is on TypeScript 4.6.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.