Fix radio button focus ring scrolling bug (#1155)
## 🤨 Rationale

Fixes #1143

## 👩‍💻 Implementation

The ring is a pseudo-element being positioned absolutely, so that it can
overlap the rest of the control. According to the docs for the
`absolute` value:

> An element with `position: absolute;` is positioned relative to the
nearest positioned ancestor (instead of positioned relative to the
viewport, like fixed).
> However; if an absolute positioned element has no positioned
ancestors, it uses the document body, and moves along with page
scrolling.

It seems that our focus ring had no positioned ancestors and was being
positioned relative to the document body. The scrollable area is
_contained_ by the document body, so scrolling it would not change the
position of the focus ring.

Note that there must have been a positioned ancestor within the
scrollable area in the radio button Storybook, because that did not
exhibit the odd behavior. The focus ring scrolled along with the
controls.

To fix this, all we have to do is set `position: relative;` on the
`.control` element which contains the focus ring pseudo-element. This
will ensure the absolute positioning is always relative to `.control`.

## 🧪 Testing

Tested in Angular example app. I also tested the other controls to make
sure nothing else was showing the same non-scrolling behavior.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.