Run Firefox tests on CI and skip existing failing tests (#1156)
## 🤨 Rationale

Relates to #1075

I fixed 7 of the 17 tests that were failing on Firefox. We want to start
running the Firefox tests on the CI (and `npm run test`), but skip the
ones known to be failing.

## 👩‍💻 Implementation

Malcolm pointed me to a package called `karma-jasmine-spec-tags` which
they had used on the WebVI team. This package allows running or skipping
tests based on arbitrary tags found in the test names. I've added
"#SkipFirefox" to the 10 remaining tests that are still failing on
Firefox. I then added `--skip-tags SkipFirefox` to the arguments passed
when executing `npm run test-firefox`. I added similar skipping support
for Chrome and Webkit.

I documented our policy for skipping tests in the CONTRIBUTING.

## 🧪 Testing

Tested that the tagged tests are indeed skipped now, and the rest are
being run on the CI.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>