Revert icon timing workaround (#1180)
# Pull Request

## 🤨 Rationale

As part of #1156 a change was made to tooltip icon tests to run in
Firefox. The implemented solution added arbitrary additional waits which
are turning out to be intermittent. This PR disables those tests until a
more robust implementation can be found. Additional investigation is
being skipped to get main stable at the moment.

Fixes #1166 (fixes the intermittency on main, and #1075 covers fixing
the actual disabled tests)

I'm also disabling the intermittent test associated with
https://github.com/ni/nimble/issues/1172 as it was causing failues
locally trying to reproduce the intermittency of this test.

## 👩‍💻 Implementation

Revert the test changes and place them into the existing original bucket
of firefox failing tests.

## 🧪 Testing

Relying on CI

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.