Enable multi-select in the table (#1181)
# Pull Request

## 🤨 Rationale

Adds multi-select support in the table as part of #856 

Functionality included in this change:
- Add selection checkboxes to rows, group rows, and table header when
multi-select is enabled

Functionality explicitly excluded from this change:
- Angular/Blazor support
- Add labels to selection checkboxes. I've added a note in the table
accessibility issue that we need to address this in a similar manner to
addressing it for the expand/collapse button on group rows: #1151
- Add support for multi-selection through CTRL+CLICK and SHIFT+CLICK

## 👩‍💻 Implementation

- Extend `TableRowSelectionMode` to include `multiple`
- Extend `TableRowSelectionState` to include `single`
- Add checkboxes to rows, group rows, and table when the selection mode
was `multiple`.
- I had to wrap the checkboxes in a `gridcell` because a checkbox isn't
an allowable direct child of a `row`. A button also isn't an allowable
direct child of a `row`, so I made the same change to wrap the existing
expand/collapse button on group rows in a `gridcell`
- I ran into issues because the `change` event is fired from the
`nimble-checkbox` when a programmatic change to the value is made.
Therefore, I am programmatically controlling the state of the selection
checkboxes rather than binding to the `checked` and `indeterminate`
properties from templates.
- Some of the TanStack selection code does not work the way we need it
to, particularly around selection + grouping. Therefore, I've written
explicit code to:
- deselect all children of a group when the group row's selection
checkbox becomes unchecked
    - determine the selection state of group rows

## 🧪 Testing

- Added a number of new unit tests
- Manually tested in storybook
- Extended table matrix tests to include all three selection modes
- Added selection modes to the column stories to enable testing
selection with other features such as column sizing and grouping

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.