Table row selection angular (#1186)
# Pull Request

## 🤨 Rationale

Add Angular support for row selection in the table, which is part of
#856.

## 👩‍💻 Implementation

- Add `selectionMode` to the table's directive
- Add `getSelectedRowIds()` and `setSelectedRowIds()` to the table's
directive
- Update the Angular example app to configure a selection mode of
`multiple`
- Update the Angular example app to render with 0 rows in the table
because there are Lighthouse failures otherwise
- This is not an ideal solution (nor a long-term solution). However, we
decided it did not make sense to block adding Angular support for row
selection on Lighthouse failures, particularly when we already have an
issue to track accessibility issues with the table -- #1151

## 🧪 Testing

- Added/ran unit tests
- Verified table worked correctly in example app

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.