Fix six tooltip tests that were failing on Firefox (#1188)
## 🤨 Rationale

See #1075 
Six tooltip tests listed in that bug were failing on Firefox, so we
marked them to be skipped.

## 👩‍💻 Implementation

The tests were checking for the display property of icons before the
anchored region of the tooltip had been loaded. This was resulting in
`getComputedStyle` for the icon to return an empty set of style info.
Waiting for the `loaded` event of the anchored region results in the
tests passing.

## 🧪 Testing

Ran tests locally

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.