Anchor tree item selection should set group selection state (#1169)
## 🤨 Rationale

There was a bug/missing feature in `nimble-anchor-tree-item`: when a
tree item is marked selected, the root level ancestor is supposed to get
a green vertical line, indicating that something within that group is
selected.


![image](https://user-images.githubusercontent.com/7282195/230658370-a98ca359-d927-4f02-bb9d-d092da6928fa.png)


## 👩‍💻 Implementation

Previously the TreeItem would listen for the `selected-change` event and update the group selection indicator based on that. It did not handle removing or adding tree items, though. I've replaced that logic with some based on a FAST `children` directive on the TreeView. This required forking the tree view's template.

## 🧪 Testing

Existing tests pass, and a couple of the tests for TreeItem were
duplicated to cover the AnchorTreeItem, too.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>