Fix and re-enable select test that was failing sporadically (#1185)
## 🤨 Rationale

Fixes #1172 
This test was disabled due to failing with a relatively high frequency.
We would like to re-enable it without causing instability.

## 👩‍💻 Implementation

I was finally able to reproduce this locally by limiting the window
dimensions sufficiently and forcing the select's listbox to open in the
"above" position. The root issue is that when this test runs late enough
in the testing sequence, there is a large matrix of dots (representing
the run tests) that Karma renders above the test control. When our
select is low on the screen, the listbox appears above the control
(since there is more window space above the control). I believe the
control renders slightly out of view, renders its listbox to fit the
remaining space, then karma scrolls the control fully into view, thereby
scrolling the listbox slightly out of view. Then our test assertion
fails because it's only 98% visible.

We can avoid this by changing the test fixture to render the component
under test in the upper left corner, on top of the report contents.

## 🧪 Testing

Ran test 

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.