Re-enable Chromatic snapshots for spinner by disabling animation (#1207)
# Pull Request

## 🤨 Rationale

Hopefully fixes #983.

We previously disabled the spinner chromatic tests due to intermittency
because Chromatic didn't reliably pause animations before taking
snapshots.

Chromatic got back to us with [more ideas for
workarounds](https://www.chromatic.com/docs/ignoring-elements#ignore-dom-elements).
While none of those ideas panned out, they inspired me to try an
approach where we explicitly disable animations on the spinner when it's
in Chromatic.

## 👩‍💻 Implementation

1. Change spinner styles to respect the a private CSS variable that
controls the `animation-play-state` attribute. T
3. Set animation play state to paused in spinner stories when they are
being rendered within Chromatic but not Storybook. This required pulling
in a copy of [the `isChromatic()`
helper](https://www.chromatic.com/docs/ischromatic). We opted to
duplicate it rather than depend on the `chromatic` package because it
has an overly large number of dependencies.

## 🧪 Testing

Locally verified that Storybook still animates the spinner and that
`animation-play-state: paused` works when I apply it manually. We'll
have to wait for Chromatic builds to tell if `isChromatic()` is working
and if this makes the test more reliable.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.